+++ .github/utils/pyproject_to_requirements.py
@@ -1,5 +1,4 @@
 import argparse
-import sys
 from pathlib import Path
 import toml